Move back to PathForBounds and move scaling to Controls
Clipping will need the bounds specified when asking for a Path.  This reworks the Shape in Controls to move the scaling of the path up into Shape.  The scaling here will check the requested bounds compared to the view's bounds and scale accordingly (taking into account Stretch/Aspect).
